refactor(forms): rename `DateOfBirthForm` and move back to common module
turns out this form is not specific to any flow, so it can be reused.
